+++ demoproject/demo_project/urls.py
@@ -6,7 +6,7 @@ from django.contrib.auth.views import login,logout
 from settings import PROJECT_ROOTDIR
 urlpatterns = patterns('',
     # Examples:    
-    url(r'^reusable-main/', include('app_folder_name.urls')),
+    url(r'^crudstuff/', include('crudstuff.urls')),
     
     
     url(r'^login/', login),