+++ docs/changelog.rst
@@ -1,7 +1,16 @@
 .. _changelog:
 
-ChangeLog for APPName
+ChangeLog for crudstuff
 =========
+
+
+0.2.0
+-----
+* Restructure and Implemented context processor based view
+* Using bindmodels class to bind models, apps, to their forms
+* Bindmodels read from settings for pluggable app 
+* Version bump
+
 
 0.1.2
 -----
@@ -15,6 +24,7 @@ ChangeLog for APPName
 ------
 * implemented login page
 * index view of customer data
+
 
 0.1
 ----